bring up to date

+++ STM32F1/cores/maple/itoa.c
@@ -124,7 +124,7 @@ extern char* ltoa( long value, char *string, int radix )
    (__GNUC_MINOR__ == 9 && __GNUC_PATCHLEVEL__ > 2)))
 extern char* utoa( unsigned value, char *string, int radix )
 #else
-extern char* utoa( unsigned long value, char *string, int radix )
+extern char* utoa( unsigned int value, char *string, int radix )
 #endif
 {
   return ultoa( value, string, radix ) ;
